chore: Move upload functions into email upload middleware
