+++ backend/src/email/middlewares/email-template.middleware.ts
@@ -1,18 +1,9 @@
 import { Request, Response, NextFunction } from 'express'
-import logger from '@core/logger'
-import {
-  MissingTemplateKeysError,
-  HydrationError,
-  RecipientColumnMissing,
-  InvalidRecipientError,
-  UnexpectedDoubleQuoteError,
-} from '@core/errors'
+import { HydrationError } from '@core/errors'
 import { TemplateError } from 'postman-templating'
-import { UploadService, StatsService, ProtectedService } from '@core/services'
-import { EmailTemplateService, EmailService } from '@email/services'
-import S3Client from '@core/services/s3-client.class'
+import { StatsService } from '@core/services'
+import { EmailTemplateService } from '@email/services'
 import { StoreTemplateOutput } from '@email/interfaces'
-import { Campaign } from '@core/models'
 
 /**
  * Store template subject and body in email template table.
@@ -77,237 +68,6 @@ const storeTemplate = async (
   }
 }
 
-/**
- * Downloads the file from s3 and checks that its columns match the attributes provided in the template.
- * If a template has not yet been uploaded, do not write to the message logs, but prompt the user to upload a template first.
- * If the template and csv do not match, prompt the user to upload a new file.
- * @param req
- * @param res
- * @param next
- */
-const uploadCompleteHandler = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<Response | void> => {
-  try {
-    const { campaignId } = req.params
-
-    // extract s3Key from transactionId
-    const { transaction_id: transactionId, filename } = req.body
-    const { s3Key } = UploadService.extractParamsFromJwt(transactionId)
-
-    // check if template exists
-    const emailTemplate = await EmailTemplateService.getFilledTemplate(
-      +campaignId
-    )
-    if (emailTemplate === null) {
-      throw new Error('Template does not exist, please create a template')
-    }
-
-    // - download from s3
-    const s3Client = new S3Client()
-    const fileContent = await s3Client.getCsvFile(s3Key)
-
-    const records = UploadService.getRecordsFromCsv(
-      +campaignId,
-      fileContent,
-      emailTemplate.params as string[]
-    )
-
-    EmailTemplateService.testHydration(
-      records,
-      emailTemplate.body as string,
-      emailTemplate.subject as string
-    )
-
-    if (EmailTemplateService.hasInvalidEmailRecipient(records)) {
-      throw new InvalidRecipientError()
-    }
-
-    // Store temp filename
-    await UploadService.storeS3TempFilename(+campaignId, filename)
-
-    try {
-      // Return early because bulk insert is slow
-      res.sendStatus(202)
-
-      // Slow bulk insert
-      await EmailService.updateCampaignAndMessages(
-        +campaignId,
-        s3Key,
-        filename,
-        records
-      )
-    } catch (err) {
-      // Do not return any response since it has already been sent
-      logger.error(
-        `Error storing messages for campaign ${campaignId}. ${err.stack}`
-      )
-      // Store error to return on poll
-      UploadService.storeS3Error(+campaignId, err.message)
-    }
-  } catch (err) {
-    const userErrors = [
-      RecipientColumnMissing,
-      MissingTemplateKeysError,
-      InvalidRecipientError,
-      UnexpectedDoubleQuoteError,
-    ]
-
-    if (userErrors.some((errType) => err instanceof errType)) {
-      return res.status(400).json({ message: err.message })
-    }
-    return next(err)
-  }
-}
-
-/*
- * Returns status of csv processing
- */
-const pollCsvStatusHandler = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<Response | void> => {
-  try {
-    const { campaignId } = req.params
-    const {
-      isCsvProcessing,
-      filename,
-      tempFilename,
-      error,
-    } = await UploadService.getCsvStatus(+campaignId)
-
-    // If done processing, returns num recipients and preview msg
-    let numRecipients, preview
-    if (!isCsvProcessing) {
-      ;[numRecipients, preview] = await Promise.all([
-        StatsService.getNumRecipients(+campaignId),
-        EmailService.getHydratedMessage(+campaignId),
-      ])
-    }
-
-    res.json({
-      is_csv_processing: isCsvProcessing,
-      csv_filename: filename,
-      temp_csv_filename: tempFilename,
-      csv_error: error,
-      num_recipients: numRecipients,
-      preview,
-    })
-  } catch (err) {
-    next(err)
-  }
-}
-
-/*
- * Deletes csv error and temp csv name from db
- */
-const deleteCsvErrorHandler = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<Response | void> => {
-  try {
-    const { campaignId } = req.params
-    await UploadService.deleteS3TempKeys(+campaignId)
-    res.sendStatus(200)
-  } catch (e) {
-    next(e)
-  }
-}
-
-// TODO: refactor this handler with uploadCompleteHandler to share the same function
-/**
- * Downloads the file from s3 and checks that its columns match the attributes provided in the template.
- * If a template has not yet been uploaded, do not write to the message logs, but prompt the user to upload a template first.
- * If the template and csv do not match, prompt the user to upload a new file.
- * @param req
- * @param res
- * @param next
- */
-const uploadProtectedCompleteHandler = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<Response | void> => {
-  try {
-    const { campaignId } = req.params
-
-    if (!(await EmailTemplateService.getFilledTemplate(+campaignId))) {
-      throw new Error('Template does not exist, please create a template')
-    }
-
-    // extract s3Key from transactionId
-    const { transaction_id: transactionId, filename } = req.body
-    const { s3Key } = UploadService.extractParamsFromJwt(transactionId) as {
-      s3Key: string
-      uploadId: string
-    }
-
-    // - download from s3
-    const s3Client = new S3Client()
-    const fileContent = await s3Client.getCsvFile(s3Key)
-
-    const records = UploadService.getProtectedRecordsFromCsv(
-      +campaignId,
-      fileContent
-    )
-
-    if (EmailTemplateService.hasInvalidEmailRecipient(records)) {
-      throw new InvalidRecipientError()
-    }
-
-    // Store temp filename
-    await UploadService.storeS3TempFilename(+campaignId, filename)
-
-    // Slow bulk insert
-    try {
-      // Return early because bulk insert is slow
-      res.sendStatus(202)
-
-      // Slow bulk insert
-      const transaction = await Campaign.sequelize?.transaction()
-      const messagesToStore = await ProtectedService.storeProtectedMessages(
-        +campaignId,
-        records,
-        transaction
-      )
-      await EmailService.updateCampaignAndMessages(
-        +campaignId,
-        s3Key,
-        filename,
-        messagesToStore,
-        transaction
-      )
-    } catch (err) {
-      // Do not return any response since it has already been sent
-      logger.error(
-        `Error storing messages for campaign ${campaignId}. ${err.stack}`
-      )
-      // Store error to return on poll
-      UploadService.storeS3Error(+campaignId, err.message)
-    }
-  } catch (err) {
-    const userErrors = [
-      RecipientColumnMissing,
-      MissingTemplateKeysError,
-      InvalidRecipientError,
-      UnexpectedDoubleQuoteError,
-    ]
-
-    if (userErrors.some((errType) => err instanceof errType)) {
-      return res.status(400).json({ message: err.message })
-    }
-    return next(err)
-  }
-}
-
 export const EmailTemplateMiddleware = {
   storeTemplate,
-  uploadCompleteHandler,
-  pollCsvStatusHandler,
-  deleteCsvErrorHandler,
-  uploadProtectedCompleteHandler,
 }
