+++ backend/src/email/middlewares/index.ts
@@ -1,3 +1,4 @@
 export * from './email.middleware'
 export * from './email-template.middleware'
 export * from './email-stats.middleware'
+export * from './email-upload.middleware'
